weather.py is updated to have temperature as interger not float

+++ weather.py
@@ -10,7 +10,7 @@ def temperature():
     req = requests.get('http://api.openweathermap.org/data/2.5/weather?zip='+zipcode+',us&appid=6ce765045581e99ca1b61b279625072d')
     json_object = req.json()
     temp_k = float(json_object['main']['temp'])
-    temp_f = (temp_k - 273.15) * 1.8 + 32
+    temp_f = int((temp_k - 273.15) * 1.8 + 32)
     return render_template('temperature.html',temp = temp_f)
 
 @app.route('/')
